added HTTP response code to the http fetcher this currently includes a function overload accepting a custom struct added a user error to the sat tracker when an HTTP error is encountered now, the new sat tracker will terminate all fetches if it gets a response != 200 this stays dead until the module is restarted moved the new sat tracker into place as sat_tracker implemented the CSV parser for Celestrak data
