+++ plugins/contests.cc
@@ -35,7 +35,7 @@ void parse_contests(xmlNode* start_node) {
 				in_item = true;
 				if (current_node->children) {
 					parse_contests(current_node->children);
-				
+
 					//*(host_api->AaediHAM_LogDebug) << "Parsed Contest ITEM tag " << temp.title << "\n";
 					in_item = false;
 					bool found = false;
@@ -87,32 +87,15 @@ int SDLCALL fetch_contests (void* data) {
 	std::string full_cache_path = host_api->AaediHAM_ConfigGetCachePath();
 	full_cache_path += "contests.cache";
 	if (contest_feed.empty()) {
-		SDL_PathInfo fileinfo;
-		
-		if (SDL_GetPathInfo(full_cache_path.c_str(), &fileinfo)) {
-			//   if (SDL_GetPathInfo("contests.cache", &fileinfo)) {
-			SDL_Time sdl_now;
-			SDL_GetCurrentTime(&sdl_now);
-			if ((sdl_now - fileinfo.modify_time) < 10800000000000  ) { // 3 Hours in ns
-				data_size = fileinfo.size;
-				disk_file.open(full_cache_path.c_str(), (std::fstream::binary | std::fstream::in ));
-				if (disk_file.is_open()) {
-					fetch_spots = (char*)malloc(fileinfo.size+1);
-					if (fetch_spots) {
-						if (disk_file.read(fetch_spots, fileinfo.size)) {
-							fetch_spots[fileinfo.size] = '\0';
-							*(host_api->AaediHAM_LogDebug) <<"Reading Contests from WA7BNM Disk Cache via timer\n";
-							file_valid = true;
-						} else {
-							free(fetch_spots);
-							fetch_spots = 0;
-						}
-					}
-					disk_file.close();
-				}
-			}
+		std::string error_string;
+		*(host_api->AaediHAM_LogDebug) <<"Reading Contests from WA7BNM Disk Cache via timer\n";
+		// 3 hour max cache age
+		data_size = disk_cache_read (full_cache_path, (void**)&fetch_spots, 10800000000000, error_string);
+		if (data_size == 0) {
+			*(host_api->AaediHAM_LogDebug) <<"Cache Result: " << error_string << "\n";
+		} else {
+			file_valid = true;
 		}
-	
 	}
 	if (!file_valid) {
 		*(host_api->AaediHAM_LogDebug) <<"Fetching Contests from WA7BNM via timer\n";
@@ -126,11 +109,11 @@ int SDLCALL fetch_contests (void* data) {
 				disk_file.write(fetch_spots, data_size);
 				if (!disk_file.good()) {
 					*(host_api->AaediHAM_LogDebug) << "Cache write failed\n";
-				
+
 				}
 			}
 			disk_file.close();
-		
+
 		}
 		xmlDocPtr xml_tree = 0;
 		xml_tree = xmlReadMemory(fetch_spots, static_cast<int>(data_size), nullptr, nullptr, 0);
@@ -209,7 +192,7 @@ void contest_plugin::plugin_main(const aaediclock_FRect& dims) const {
 	         contest_page[0]=0;
 	    }
 	}
-	
+
 	 aaediclock_FRect TextRect;
 	 aaediclock_Color panel_color;
 	 if (cycle) {
@@ -217,9 +200,9 @@ void contest_plugin::plugin_main(const aaediclock_FRect& dims) const {
 	 } else {
 	     panel_color = {0, 128, 200, 255};
 	 }
-	
+
 	 host_api->AaediHAM_GraphicsClear();
-	
+
 	 float unitx = dims.w/20;
 	 float unity = dims.h/15;
 	 TextRect.w=unitx*18;
