+++ plugins/sat_tracker.h
@@ -1,63 +1,95 @@
-#ifndef SAT_TRACKER_H
-#define SAT_TRACKER_H
-#include "plugin_api.h"
-
-
-#include <libsgp4/SGP4.h>
-#include <libsgp4/Observer.h>
-#include <memory>
-#include "../aaediclock.h"
-#include <math.h>
+#include "aaediclock.h"
+#include <libxml/parser.h>
+#include "SGP4/SGP4.h"
 #ifndef M_PI
 #define M_PI 3.14159265358979323846
 #endif
 
-class TrackedSatellite {
 
-    private:
-        struct SatTelemetry {
-            double lat;
-            double lon;
-            double azimuth;     // compass direction from DE
-            double elevation;   // elevation above horizon relative to DE
-            time_t timestamp;
-        };
-
-        std::string name;
-        std::string tle1;
-        std::string tle2;
-        libsgp4::Tle* sat_tle;
-        libsgp4::SGP4* sgp4;
-        std::vector<struct SatTelemetry> telemetry;
-
-    public:
-        aaediclock_Color color;
-        TrackedSatellite(const std::string& source_name, const std::string& l1, const std::string& l2);
-        ~TrackedSatellite();
-        TrackedSatellite(TrackedSatellite&& source) noexcept;	// move constructor
-        TrackedSatellite& operator=(TrackedSatellite&& source) noexcept;     // move with replace
-        TrackedSatellite(const TrackedSatellite& source);		// copy to new
-        TrackedSatellite& operator=(const TrackedSatellite& source);	// copy over existing
-        const std::string& get_name() const;
-        void new_tracking(const std::string& source_name, const std::string& l1, const std::string& l2);
-        time_t pass_start();
-        time_t pass_end();
-        time_t telemetry_age();
-        void draw_pass(const time_t pass_start, const time_t pass_end,  std::vector<aaediclock_FPoint> *pass_pts, const aaediclock_FRect *size);
-        void add_telemetry(const double lat,const double lon, const double elevation, const double azimuth, const time_t timestamp);
-        void location (aaediclock_FPoint *result);
-        bool gen_telemetry(const int resolution, libsgp4::Observer& obs);
-        void draw_telemetry(aaediclock_FRect& dims);
+struct vector3 {
+    double x;
+    double y;
+    double z;
 };
 
-//void sat_tracker (ScreenFrame& panel, TTF_Font* font, ScreenFrame& map);
 
-class DllExport sat_tracker_plugin : public aaediclock_plugin_api {
+
+       struct SatTelemetry {
+            double lat		= 0.0;		// Terrestrial latitude
+            double lon		= 0.0;		// Terrestrial longitude
+            double alt		= 0.0;		// Topocentric altitude
+            double dec		= 0.0;		// declination in RADIANS
+            double ra		= 0.0;		// right ascention in RADIANS
+            double azimuth      = 0.0;		// compass direction from DE
+            double elevation    = 0.0;		// elevation above horizon relative to DE
+            time_t timestamp 	= 0;		// Seconds since Unix Epoch
+            double tsince 	= 0.0;		// Julian Date minutes since OMM Record Epoch
+        };
+
+
+
+
+// core Sat data struct
+
+class OMMRecord {
+    public:
+        // internal state
+        bool                valid   = false;
+        std::string         message;
+        // object identification
+        std::string         name;
+        std::string         object_id;
+        std::string         norad_id;
+        uint32_t            element_set_no;
+        char                class_type;
+        aaediclock_Color	color;
+        // Earth relative orbital details
+        std::string		epoch_raw;
+        double              epoch_jd;
+        double              mean_motion;
+        double              eccentricity;
+        double              inclination;
+        double              right_ascension_of_ascending_node;
+        double              arg_pericenter;
+        double              mean_anomaly;
+        int                 ephemeris_type;
+        uint32_t            revolution_at_epoch;
+        double              bstar;                  // atmospheric drag //  SGP4 drag term
+        double              mean_motion_dot;        // first derivative
+        double              mean_motion_ddot;       // second derivative // normally 0
+        elsetrec 		satrec = {};
+        std::vector<struct SatTelemetry> telemetry;
+        bool generate_telemetry(int resolution_min);
+        time_t telemetry_age();
+        void draw_pass(const time_t pass_start, const time_t pass_end,  std::vector<aaediclock_FPoint> *pass_pts, const aaediclock_FRect *size);
+
+        time_t pass_start();
+        time_t pass_end();
+        bool sgp4_init();
+        void location (aaediclock_FPoint *result);
+
+};
+
+
+
+// everything that parses input from Celestrak goes here
+namespace SGP4Parser {
+    namespace XML {
+        void process_node(xmlNode* start_node, OMMRecord& result);
+    }
+    double ISO8601_to_Julian(std::string);
+    bool fromXML(std::string& input);
+
+    namespace CSV {
+        void parse_headers(const std::string& input, std::vector<std::string>& headers);
+    }
+    bool fromCSV(std::string& input);
+}
+
+class DllExport new_sat_tracker_plugin : public aaediclock_plugin_api {
         void plugin_init() const override;
         void plugin_main(const aaediclock_FRect& dims) const override;
         const char* getName() const override;
         void plugin_exit() const override;
         void set_host(aaediclock_host_api* host);
 };
-
-#endif
